refactor: isolate Pascal frontend
