+++ src/clike/stubs.c
@@ -1,4 +1,4 @@
-#include "frontend/ast.h"
+#include "Pascal/ast.h"
 #include "symbol/symbol.h"
 #include "core/types.h"
 #include "core/utils.h"
