+++ src/main.c
@@ -1,6 +1,6 @@
-#include "frontend/lexer.h"
-#include "frontend/parser.h"
-#include "frontend/ast.h"
+#include "Pascal/lexer.h"
+#include "Pascal/parser.h"
+#include "Pascal/ast.h"
 #include "core/types.h"
 #include "core/utils.h"
 #include "core/list.h"
@@ -17,7 +17,7 @@
 #include <SDL2/SDL_ttf.h>
 #endif
 #include "vm/vm.h"
-// frontend/ast.h is already included via globals.h or directly, no need for duplicate
+// Pascal/ast.h is already included via globals.h or directly, no need for duplicate
 
 /* Global variables */
 int gParamCount = 0;
